PostGISRaster: implement deferred driver loading
