+++ frmts/gdalallregister.cpp
@@ -190,6 +190,9 @@ void CPL_STDCALL GDALAllRegister()
 #if defined(DEFERRED_PARQUET_DRIVER)
     DeclareDeferredOGRParquetPlugin();
 #endif
+#if defined(DEFERRED_POSTGISRASTER_DRIVER)
+    DeclareDeferredPostGISRasterPlugin();
+#endif
 #if defined(DEFERRED_PDF_DRIVER)
     DeclareDeferredPDFPlugin();
 #endif
