+++ frmts/postgisraster/postgisrasterdataset.cpp
@@ -36,6 +36,7 @@
 
 #include "gdal_frmts.h"
 #include "postgisraster.h"
+#include "postgisrasterdrivercore.h"
 #include <math.h>
 
 #include <algorithm>
@@ -2801,7 +2802,7 @@ GetConnectionInfo(const char *pszFilename, char **ppszConnectionString,
 {
     int nPos = -1, sPos = -1, i;
     char *pszTmp = nullptr;
-    char **papszParams = ParseConnectionString(pszFilename);
+    char **papszParams = PostGISRasterParseConnectionString(pszFilename);
     if (papszParams == nullptr)
     {
         return false;
@@ -3146,11 +3147,9 @@ static PGconn *GetConnection(const char *pszFilename,
         /**************************************************************
          * Open a new database connection
          **************************************************************/
-        PostGISRasterDriver *poDriver = static_cast<PostGISRasterDriver *>(
-            GDALGetDriverByName("PostGISRaster"));
-
-        poConn = poDriver->GetConnection(*ppszConnectionString, pszService,
-                                         pszDbname, pszHost, pszPort, pszUser);
+        poConn = PostGISRasterDriver::gpoPostGISRasterDriver->GetConnection(
+            *ppszConnectionString, pszService, pszDbname, pszHost, pszPort,
+            pszUser);
 
         if (poConn == nullptr)
         {
@@ -3167,29 +3166,6 @@ static PGconn *GetConnection(const char *pszFilename,
     CPLFree(pszPassword);
 
     return poConn;
-}
-
-/************************************************************************/
-/*                            Identify()                                */
-/************************************************************************/
-
-int PostGISRasterDataset::Identify(GDALOpenInfo *poOpenInfo)
-{
-    if (poOpenInfo->pszFilename == nullptr || poOpenInfo->fpL != nullptr ||
-        !STARTS_WITH_CI(poOpenInfo->pszFilename, "PG:"))
-    {
-        return FALSE;
-    }
-
-    // Will avoid a OGR PostgreSQL connection string to be recognized as a
-    // PostgisRaster one and later fail (#6034)
-    if (strstr(poOpenInfo->pszFilename, " schemas=") ||
-        strstr(poOpenInfo->pszFilename, " SCHEMAS="))
-    {
-        return FALSE;
-    }
-
-    return TRUE;
 }
 
 /***********************************************************************
@@ -3219,7 +3195,7 @@ GDALDataset *PostGISRasterDataset::Open(GDALOpenInfo *poOpenInfo)
     /**************************
      * Check input parameter
      **************************/
-    if (!Identify(poOpenInfo))
+    if (!PostGISRasterDriverIdentify(poOpenInfo))
         return nullptr;
 
     poConn = GetConnection(poOpenInfo->pszFilename, &pszConnectionString,
@@ -4092,85 +4068,6 @@ GBool PostGISRasterDataset::PolygonFromCoords(int nXOff, int nYOff,
     return true;
 }
 
-/************************************************************************/
-/*                    PostGISRasterDriverGetSubdatasetInfo()            */
-/************************************************************************/
-
-struct PostGISRasterDriverSubdatasetInfo : public GDALSubdatasetInfo
-{
-  public:
-    explicit PostGISRasterDriverSubdatasetInfo(const std::string &fileName)
-        : GDALSubdatasetInfo(fileName)
-    {
-    }
-
-    // GDALSubdatasetInfo interface
-  private:
-    void parseFileName() override
-    {
-        if (!STARTS_WITH_CI(m_fileName.c_str(), "PG:"))
-        {
-            return;
-        }
-
-        char **papszParams = ParseConnectionString(m_fileName.c_str());
-
-        const int nTableIdx = CSLFindName(papszParams, "table");
-        if (nTableIdx != -1)
-        {
-            size_t nTableStart = m_fileName.find("table=");
-            bool bHasQuotes{false};
-            try
-            {
-                bHasQuotes = m_fileName.at(nTableStart + 6) == '\'';
-            }
-            catch (const std::out_of_range &)
-            {
-                // ignore error
-            }
-
-            m_subdatasetComponent = papszParams[nTableIdx];
-
-            if (bHasQuotes)
-            {
-                m_subdatasetComponent.insert(6, "'");
-                m_subdatasetComponent.push_back('\'');
-            }
-
-            m_driverPrefixComponent = "PG";
-
-            size_t nPathLength = m_subdatasetComponent.length();
-            if (nTableStart != 0)
-            {
-                nPathLength++;
-                nTableStart--;
-            }
-
-            m_pathComponent = m_fileName;
-            m_pathComponent.erase(nTableStart, nPathLength);
-            m_pathComponent.erase(0, 3);
-        }
-
-        CSLDestroy(papszParams);
-    }
-};
-
-static GDALSubdatasetInfo *
-PostGISRasterDriverGetSubdatasetInfo(const char *pszFileName)
-{
-    if (STARTS_WITH_CI(pszFileName, "PG:"))
-    {
-        std::unique_ptr<GDALSubdatasetInfo> info =
-            std::make_unique<PostGISRasterDriverSubdatasetInfo>(pszFileName);
-        if (!info->GetSubdatasetComponent().empty() &&
-            !info->GetPathComponent().empty())
-        {
-            return info.release();
-        }
-    }
-    return nullptr;
-}
-
 /***********************************************************************
  * GDALRegister_PostGISRaster()
  **********************************************************************/
@@ -4180,21 +4077,15 @@ void GDALRegister_PostGISRaster()
     if (!GDAL_CHECK_VERSION("PostGISRaster driver"))
         return;
 
-    if (GDALGetDriverByName("PostGISRaster") != nullptr)
+    if (GDALGetDriverByName(DRIVER_NAME) != nullptr)
         return;
 
     GDALDriver *poDriver = new PostGISRasterDriver();
-
-    poDriver->SetDescription("PostGISRaster");
-    poDriver->SetMetadataItem(GDAL_DCAP_RASTER, "YES");
-    poDriver->SetMetadataItem(GDAL_DMD_LONGNAME, "PostGIS Raster driver");
-    poDriver->SetMetadataItem(GDAL_DMD_SUBDATASETS, "YES");
+    PostGISRasterDriverSetCommonMetadata(poDriver);
 
     poDriver->pfnOpen = PostGISRasterDataset::Open;
-    poDriver->pfnIdentify = PostGISRasterDataset::Identify;
     poDriver->pfnCreateCopy = PostGISRasterDataset::CreateCopy;
     poDriver->pfnDelete = PostGISRasterDataset::Delete;
-    poDriver->pfnGetSubdatasetInfoFunc = PostGISRasterDriverGetSubdatasetInfo;
 
     GetGDALDriverManager()->RegisterDriver(poDriver);
 }
